Fixed bugs in QUESTFIT (conversion to Flambda is now only applied to templates & before they're normalised), created Jupyter notebook for testing fitting the MIR lines of a Spitzer spectrum: run_q3dfit_MIRlines.ipynb
Note on units:  In QUESTFIT, Flambda is now fitted in units of [1e-10 erg s^-1 cm^-2 micron^-1];  the normalisation factor was introduced as lmfit did not recover a good fit if there was a large difference in scaling between the spectrum and templates. (Templates are normalised to peak at a value of 1).